% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_script_description.R
\name{insert_script_description}
\alias{insert_script_description}
\title{Insert script description}
\usage{
insert_script_description(script_title = "Script title",
  project_title = "My research project",
  script_type = "Data processing", script_description = "Description",
  script_date = Rahat::today(format = "human"),
  creator_name = "Mirza Cengic", creator_email = "mirzaceng@gmail.com")
}
\arguments{
\item{script_title}{Character. Script title.}

\item{project_title}{Character. To which project does script relate to.}

\item{script_type}{Character. Type of script. Eg. data analysis, modeling, plotting.}

\item{script_description}{Character. What it does, and why, in more detial.}

\item{script_date}{Character. Default is Day, Month Year.}

\item{creator_name}{Character. Default is "Mirza Cengic".}

\item{creator_email}{Character. Default is "mirzaceng@gmail.com"}
}
\value{
None.
}
\description{
Use this function to insert template used to give some basic information on the current R script to its top.
This function will only work in RStudio, as it uses \code{rstudioapi} package to add text.
}
\examples{
insert_slurm_header()
}
