% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_slurm_header.R
\name{insert_slurm_header}
\alias{insert_slurm_header}
\title{Insert SLURM header}
\usage{
insert_slurm_header(partition = "milkun", email_type = "FAIL",
  email = "mirzaceng@gmail.com", memory = "20G", time = "6:00:00",
  log_file = "")
}
\arguments{
\item{partition}{Character. Cluster partition. Default is milkun. Use either milkun or milkunshort.}

\item{email_type}{Character. Type of email to send to user. See SLURM help. Default is FAIL.}

\item{email}{Character. Email address to which send the emails to. Default is mirzaceng@gmail.com.}

\item{memory}{Character. Memory allocated to the process. Default is 20G. Add the required unit to the memory amount.}

\item{time}{Character. Time allocated to the process. Default is 6 hours (6:00:00).}

\item{log_file}{Character. Path for the log file. Default path is empty. Example: \code{log_file = "/vol/milkun/my_log.log"}}
}
\value{
None.
}
\description{
Use this function to insert SLURM header to the top of the current R script.
This function will only work in RStudio, as it uses \code{rstudioapi} package to add text. Default values are higly customized for me!
}
\examples{
insert_slurm_header()
}
