% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milkunize.R
\name{milkunize}
\alias{milkunize}
\title{Generate file path for the milkun drives}
\usage{
milkunize(path, drive = "m1")
}
\arguments{
\item{path}{Folder path}

\item{drive}{Milkun drive.}
}
\value{
Folder path with milkun location
}
\description{
Generate file path for the milkun drives
}
\examples{
milkunize("Projects/Land_use")
}
