% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milkunize.R
\name{milkunize2}
\alias{milkunize2}
\title{Milkunize2}
\usage{
milkunize2(path, drive = "home", method = "user")
}
\arguments{
\item{path}{Path that will be milkunized}

\item{drive}{Milkun drive. One in c("home", "archive", "data").}

\item{method}{Method how the local system is detected. Either sysname (windows or linux), or user (mcengic on cluster or mirzacengic on the laptop).}
}
\value{
Folder path with milkun2 location
}
\description{
Generate file path for the milkun drives. Version 2 of the function is for updated milkun drives (milkundata, milkunB, and milkunarc)
}
\examples{
milkunize("Projects/Land_use")
}
