% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csvs.R
\name{read_csvs}
\alias{read_csvs}
\title{Read multiple csv files}
\usage{
read_csvs(folder_path, pattern = "*.csv$")
}
\arguments{
\item{folder_path}{Folder path where the files will be loaded from.}

\item{pattern}{Pattern to match files. Default is "*.csv"}
}
\value{
Multiple dataframes
}
\description{
This function is \emph{deprecated}. Use purrr with \code{map(read_csv); reduce(rbind)}approach.
This function will load all the csv files from the given directory with the selected pattern.
}
\examples{
None.
}
