% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today.R
\name{today}
\alias{today}
\title{Today
Get today's date}
\usage{
today(sep = "_")
}
\arguments{
\item{sep}{Format separator. Character. Default is "_".}
}
\value{
Character. Day Month Year format.
}
\description{
Today
Get today's date
}
\examples{
today()
}
