% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milkunize.R
\name{milkunize}
\alias{milkunize}
\title{Milkunize}
\usage{
milkunize(path, drive = "m1", method = "user")
}
\arguments{
\item{path}{Folder path}

\item{drive}{Milkun drive. One in c("m1", "m2", "m5", "m17").}

\item{method}{Method how the local system is detected. Either sysname (windows or linux), or user (mcengic on cluster or mirzacengic on the laptop).}
}
\value{
Folder path with milkun location
}
\description{
Generate file path for the milkun drives
}
\examples{
milkunize("Projects/Land_use")
}
