% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/today.R
\name{today}
\alias{today}
\title{Today
Get today's date}
\usage{
today(format = "computer", sep = "_")
}
\arguments{
\item{format}{Character. Return computer-friendly or human-friendly format. One in c("computer", "human").}

\item{sep}{Character. Format separator. Default is "_".}
}
\value{
Character.
}
\description{
Today
Get today's date
}
\examples{
today()
}
